% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Get species occurence data}
\usage{
getdata(grid, g, latlong)
}
\arguments{
\item{grid}{grid polygons}

\item{g}{index of the grid polygon}

\item{latlong}{Latitude Longitude coordinate reference system}
}
\value{

}
\description{
Get species occurence data
}
\examples{
occ <- getdata(grid, g,latlong)
}
