% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isaquatic.R
\name{isaquatic}
\alias{isaquatic}
\title{Are species occurences aquatic?}
\usage{
isaquatic(occ, wet)
}
\arguments{
\item{occ}{Species occurence dataframe}

\item{wet}{'wet' polygons}
}
\value{

}
\description{
Are species occurences aquatic?
}
\examples{
isaquatic(occ,wet)
}
